mobile: match current gpui Platform trait (#2)
gpui-mobile was built against zed rev 5688167d22; current Zed HEAD has
moved on. Patch the iOS and Android Platform / PlatformDisplay impls
to satisfy the current trait surface:

* Implement Platform::hide_cursor_until_mouse_moves and
  Platform::is_cursor_visible on iOS, Android, and the Android
  SharedPlatform wrapper. All are no-ops / false on mobile, which has
  no system mouse cursor.
* DisplayId switched from u32 to u64. Pass the raw u64 pointer-derived
  id through directly instead of truncating.

Testing:
  cargo check -p gpui_mobile --target aarch64-apple-ios            # clean
  cargo check -p gpui_mobile --target aarch64-linux-android        # clean
  (Android build needs the NDK r29 toolchain and these env vars:
     RUST_FONTCONFIG_DLOPEN=on
     CC_aarch64_linux_android=<NDK>/aarch64-linux-android26-clang
     AR_aarch64_linux_android=<NDK>/llvm-ar
     CARGO_TARGET_AARCH64_LINUX_ANDROID_LINKER=<NDK>/aarch64-linux-android26-clang)

Upstream: not yet proposed

+++ crates/gpui_mobile/src/ios/display.rs
@@ -66,8 +66,8 @@ impl IosDisplay {
 
 impl PlatformDisplay for IosDisplay {
     fn id(&self) -> DisplayId {
-        // iOS doesn't have display IDs like macOS, so we use the screen pointer as an ID
-        DisplayId::new(self.screen as u32)
+        // iOS doesn't have display IDs like macOS, so we use the screen pointer as an ID.
+        DisplayId::new(self.screen as u64)
     }
 
     fn uuid(&self) -> Result<Uuid> {
